Add repeat count setting for verse playback
Introduces a new spinbox in the settings dialog to allow users to set the number of times a verse is repeated during playback. Updates the ListeningSettings class to include a repeat_count attribute with a default value of 1, and ensures the value is saved and loaded with the rest of the settings.